requirements.txt
gradio>=4.0.0
torch>=2.0.0
transformers>=4.30.0
openai-whisper>=20230314
gtts>=2.3.0
accelerate>=0.20.0